superadmin/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy[asyncio]==2.0.35
asyncpg==0.30.0
alembic==1.13.3
pydantic==2.9.2
pydantic-settings==2.5.2
pyjwt[crypto]==2.9.0
bcrypt==4.2.0
httpx==0.27.2          # For calling deployment APIs
python-dotenv==1.0.1
